test(#1664): relocate clone-count regression guard out of the 13k-line merge/mod.rs
File-size hygiene follow-up: move the merger_over_runs helper and the
kway_merge_does_not_double_clone_entries regression test from
merge/mod.rs into a new sibling file merge/clone_regression_tests.rs
(#cfg(test) child module of merge, same pattern as teardown_tests.rs),
so mod.rs's net diff shrinks to the two production fix edits + the
`peek` #[cfg(test)] gate + a one-line mod declaration.

No behavior change: same test body, same documented main-vs-post-fix
clone counts (30 -> 0 for N=15), verified green in its new location.
